chore: refactor for testability
